Issue #220:Move deleteBooking from AmenityController to BookingController
