+++ service/src/main/java/com/myhome/controllers/BookingController.java
@@ -1,0 +1,29 @@
+package com.myhome.controllers;
+
+import com.myhome.api.BookingsApi;
+import com.myhome.services.BookingService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@Slf4j
+@RequiredArgsConstructor
+public class BookingController implements BookingsApi {
+
+  private final BookingService bookingSDJpaService;
+
+  @Override
+  public ResponseEntity<Void> deleteBooking(@PathVariable String amenityId,
+      @PathVariable String bookingId) {
+    boolean isBookingDeleted = bookingSDJpaService.deleteBooking(bookingId);
+    if (isBookingDeleted) {
+      return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    } else {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+    }
+  }
+}
