+++ service/src/main/java/com/myhome/services/AmenityService.java
@@ -33,6 +33,4 @@ public interface AmenityService {
   Set<Amenity> listAllAmenities(String communityId);
 
   boolean updateAmenity(AmenityDto updatedAmenityDto);
-
-  boolean deleteBooking(String bookingId);
 }
