+++ service/src/main/java/com/myhome/services/BookingService.java
@@ -1,0 +1,7 @@
+package com.myhome.services;
+
+public interface BookingService {
+
+  boolean deleteBooking(String bookingId);
+
+}
