+++ service/src/main/java/com/myhome/services/springdatajpa/AmenitySDJpaService.java
@@ -20,7 +20,6 @@ import com.myhome.controllers.mapper.AmenityApiMapper;
 import com.myhome.domain.Amenity;
 import com.myhome.domain.Community;
 import com.myhome.model.AmenityDto;
-import com.myhome.repositories.AmenityBookingItemRepository;
 import com.myhome.repositories.AmenityRepository;
 import com.myhome.repositories.CommunityRepository;
 import com.myhome.services.AmenityService;
@@ -41,7 +40,6 @@ public class AmenitySDJpaService implements AmenityService {
   private final CommunityRepository communityRepository;
   private final CommunityService communityService;
   private final AmenityApiMapper amenityApiMapper;
-  private final AmenityBookingItemRepository bookingRepository;
 
   @Override
   public Optional<List<AmenityDto>> createAmenities(Set<AmenityDto> amenities, String communityId) {
@@ -103,15 +101,5 @@ public class AmenitySDJpaService implements AmenityService {
             })
             .orElse(null))
         .map(amenityRepository::save).isPresent();
-  }
-
-  @Override
-  public boolean deleteBooking(String bookingId) {
-    return bookingRepository.findByAmenityBookingItemId(bookingId)
-        .map(bookingItem -> {
-          bookingRepository.delete(bookingItem);
-          return true;
-        })
-        .orElse(false);
   }
 }
