+++ service/src/test/java/com/myhome/services/unit/AmenitySDJpaServiceTest.java
@@ -18,10 +18,8 @@ package com.myhome.services.unit;
 
 import com.myhome.controllers.mapper.AmenityApiMapper;
 import com.myhome.domain.Amenity;
-import com.myhome.domain.AmenityBookingItem;
 import com.myhome.domain.Community;
 import com.myhome.model.AmenityDto;
-import com.myhome.repositories.AmenityBookingItemRepository;
 import com.myhome.repositories.AmenityRepository;
 import com.myhome.repositories.CommunityRepository;
 import com.myhome.services.CommunityService;
@@ -54,7 +52,6 @@ class AmenitySDJpaServiceTest {
 
   private static final String TEST_AMENITY_NAME = "test-amenity-name";
   private static final BigDecimal TEST_AMENITY_PRICE = BigDecimal.valueOf(1);
-  private static final String TEST_BOOKING_ID = "test-booking-id";
   private final String TEST_AMENITY_ID = "test-amenity-id";
   private final String TEST_AMENITY_DESCRIPTION = "test-amenity-description";
   private final String TEST_COMMUNITY_ID = "test-community-id";
@@ -67,8 +64,6 @@ class AmenitySDJpaServiceTest {
   private CommunityService communityService;
   @Mock
   private AmenityApiMapper amenityApiMapper;
-  @Mock
-  private AmenityBookingItemRepository bookingItemRepository;
 
   @InjectMocks
   private AmenitySDJpaService amenitySDJpaService;
@@ -296,38 +291,6 @@ class AmenitySDJpaServiceTest {
     verifyNoMoreInteractions(amenityRepository);
   }
 
-  @Test
-  void deleteBookingItem() {
-    // given
-    AmenityBookingItem testBookingItem = getTestBookingItem();
-
-    given(bookingItemRepository.findByAmenityBookingItemId(TEST_BOOKING_ID))
-        .willReturn(Optional.of(testBookingItem));
-
-    // when
-    boolean bookingDeleted = amenitySDJpaService.deleteBooking(TEST_BOOKING_ID);
-
-    // then
-    assertTrue(bookingDeleted);
-    verify(bookingItemRepository).findByAmenityBookingItemId(TEST_BOOKING_ID);
-    verify(bookingItemRepository).delete(testBookingItem);
-  }
-
-  @Test
-  void deleteBookingNotExists() {
-    // given
-    given(bookingItemRepository.findByAmenityBookingItemId(TEST_BOOKING_ID))
-        .willReturn(Optional.empty());
-
-    // when
-    boolean amenityDeleted = amenitySDJpaService.deleteBooking(TEST_BOOKING_ID);
-
-    // then
-    assertFalse(amenityDeleted);
-    verify(bookingItemRepository).findByAmenityBookingItemId(TEST_BOOKING_ID);
-    verify(bookingItemRepository, never()).delete(any());
-  }
-
   private AmenityDto getTestAmenityDto() {
     Long TEST_AMENITY_ENTITY_ID = 1L;
 
@@ -348,10 +311,5 @@ class AmenitySDJpaServiceTest {
         .withPrice(communityAmenityDto.getPrice())
         .withDescription(communityAmenityDto.getDescription())
         .withCommunity(TestUtils.CommunityHelpers.getTestCommunity());
-  }
-
-  private AmenityBookingItem getTestBookingItem() {
-    return new AmenityBookingItem()
-        .withAmenityBookingItemId(TEST_BOOKING_ID);
   }
 }
